// src/app/api/model/getLoansLoanIdTransactionsTransactionIdResponse.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetLoanTransactionRelation } from './getLoanTransactionRelation';
import { GetLoansCurrency } from './getLoansCurrency';
import { GetLoansLoanIdLoanChargePaidByData } from './getLoansLoanIdLoanChargePaidByData';
import { GetCodeValuesDataResponse } from './getCodeValuesDataResponse';
import { PaymentDetailData } from './paymentDetailData';
import { GetLoansType } from './getLoansType';


/**
 * GetLoansLoanIdTransactionsTransactionIdResponse
 */
export interface GetLoansLoanIdTransactionsTransactionIdResponse { 
    amount?: number;
    classification?: GetCodeValuesDataResponse;
    currency?: GetLoansCurrency;
    date?: string;
    externalId?: string;
    feeChargesPortion?: number;
    id?: number;
    interestPortion?: number;
    loanChargePaidByList?: Set<GetLoansLoanIdLoanChargePaidByData>;
    manuallyReversed?: boolean;
    netDisbursalAmount?: number;
    outstandingLoanBalance?: number;
    overpaymentPortion?: number;
    paymentDetailData?: PaymentDetailData;
    penaltyChargesPortion?: number;
    possibleNextRepaymentDate?: string;
    principalPortion?: number;
    reversalExternalId?: string;
    reversedOnDate?: string;
    submittedOnDate?: string;
    transactionRelations?: Set<GetLoanTransactionRelation>;
    type?: GetLoansType;
    unrecognizedIncomePortion?: number;
}